Refactor a small portion of the ServeSystem.
Refactored the serve code in the ServeSystem, following the pattern of
the BuildSystem & WatchSystem.
